Convert backend data access for completely revised 0.13.0.
Also folded various backend traits into one uber trait, which is
easier to manage and (in future) make more consistent.
