+++ app/AppModule.scala
@@ -8,7 +8,7 @@ import backend.googledocs.GoogleDocsHtmlPages
 import backend.helpdesk.EhriHelpdesk
 import backend.parse.{ParseCypherQueryDAO, ParseFeedbackDAO}
 import backend.rest.cypher.{Cypher, CypherDAO}
-import backend.rest.{RestBackend, CypherIdGenerator, GidSearchResolver, SearchDAO}
+import backend.rest.{RestBackend, CypherIdGenerator, GidSearchResolver}
 import com.google.inject.AbstractModule
 import eu.ehri.project.indexing.index.Index
 import eu.ehri.project.indexing.index.impl.SolrIndex
@@ -35,7 +35,6 @@ class AppModule extends AbstractModule {
     bind(classOf[SearchItemResolver]).to(classOf[GidSearchResolver])
     bind(classOf[EventHandler]).to(classOf[GlobalEventHandler])
     bind(classOf[Backend]).to(classOf[RestBackend])
-    bind(classOf[SearchDAO]).to(classOf[backend.rest.Search])
     bind(classOf[FeedbackDAO]).to(classOf[ParseFeedbackDAO])
     bind(classOf[CypherQueryDAO]).to(classOf[ParseCypherQueryDAO])
     bind(classOf[HelpdeskDAO]).to(classOf[EhriHelpdesk])
