--- modules/backend/src/main/scala/backend/VirtualCollections.scala
@@ -1,30 +1,0 @@
-package backend
-
-import scala.concurrent.Future
-
-trait VirtualCollections {
-  /**
-   * Include a set of items in a virtual collection.
-   * 
-   * @param vcId the virtual collection id
-   * @param ids a set of item ids
-   */
-  def addReferences[MT: Resource](vcId: String, ids: Seq[String]): Future[Unit]
-
-  /**
-   * Remove a set of items from a virtual collection.
-   *
-   * @param vcId the virtual collection id
-   * @param ids a set of item ids
-   */
-  def deleteReferences[MT: Resource](vcId: String, ids: Seq[String]): Future[Unit]
-
-  /**
-   * Move a set of items from one virtual collection to another
-   *
-   * @param fromVc the source virtual collection id
-   * @param toVc the destination virtual collection id
-   * @param ids a set of item ids
-   */
-  def moveReferences[MT: Resource](fromVc: String, toVc: String, ids: Seq[String]): Future[Unit]
-}
