--- modules/backend/src/main/scala/backend/Visibility.scala
@@ -1,7 +1,0 @@
-package backend
-
-import scala.concurrent.Future
-
-trait Visibility {
-  def setVisibility[MT: Resource](id: String, data: Seq[String]): Future[MT]
-}
