+++ modules/backend/src/main/scala/backend/rest/RestContext.scala
@@ -1,6 +1,5 @@
 package backend.rest
 
-import play.api.cache.CacheApi
 import backend.{ApiUser, EventHandler}
 
 import scala.concurrent.ExecutionContext
