+++ modules/portal/app/controllers/portal/Helpdesk.scala
@@ -9,7 +9,6 @@ import play.api.mvc.RequestHeader
 import backend.{BadHelpdeskResponse, HelpdeskDAO, Backend}
 import models.{UserProfile, Repository}
 import javax.inject._
-import backend.rest.SearchDAO
 import controllers.portal.base.PortalController
 import utils.MovedPageLookup
 import utils.search.SearchItemResolver
@@ -25,8 +24,7 @@ case class Helpdesk @Inject()(
   mailer: MailerClient,
   helpdeskDAO: HelpdeskDAO,
   pageRelocator: MovedPageLookup,
-  messagesApi: MessagesApi,
-  search: SearchDAO
+  messagesApi: MessagesApi
 ) extends PortalController {
 
   import play.api.data.Form
@@ -46,13 +44,13 @@ case class Helpdesk @Inject()(
 
   def helpdesk = OptionalUserAction.async { implicit request =>
     helpdeskDAO.available.flatMap { repos =>
-      search.list[Repository](repos.map(_._1)).map { institutions =>
+      userBackend.fetch[Repository](repos.map(_._1)).map { institutions =>
         Ok(views.html.helpdesk.helpdesk(prefilledForm, institutions))
       }
     }
   }
 
-  private def sendMessageEmail(emailAddress: String, query: String)(implicit request: RequestHeader): Unit = {
+  private def sendMessageEmail(emailAddress: String, query: String)(implicit request: RequestHeader) {
     val email = Email(
       subject = "EHRI Portal Helpdesk",
       to = Seq(emailAddress),
@@ -67,7 +65,7 @@ case class Helpdesk @Inject()(
     val boundForm = helpdeskForm.bindFromRequest
     boundForm.fold(
       errorForm =>helpdeskDAO.available.flatMap { repos =>
-        search.list[Repository](repos.map(_._1)).map { institutions =>
+        userBackend.fetch[Repository](repos.map(_._1)).map { institutions =>
           BadRequest(views.html.helpdesk.helpdesk(errorForm, institutions))
         }
       },
@@ -75,7 +73,7 @@ case class Helpdesk @Inject()(
         val (email, query, copyMe) = data
         if (email.isEmpty && copyMe) {
           helpdeskDAO.available.flatMap { repos =>
-            search.list[Repository](repos.map(_._1)).map { institutions =>
+            userBackend.fetch[Repository](repos.map(_._1)).map { institutions =>
               BadRequest(views.html.helpdesk.helpdesk(
                 boundForm
                   .withError("email", "error.required"), institutions))
@@ -87,7 +85,7 @@ case class Helpdesk @Inject()(
           }
 
           helpdeskDAO.askQuery(query).flatMap { responses =>
-            search.list[Repository](responses.map(_._1)).map { institutions =>
+            userBackend.fetch[Repository](responses.map(_._1)).map { institutions =>
               Ok(views.html.helpdesk.results(prefilledForm.fill(data), query, responses, institutions))
             }
           } recover {
