+++ modules/portal/app/controllers/portal/HistoricalAgents.scala
@@ -50,7 +50,7 @@ case class HistoricalAgents @Inject()(
   def export(id: String) = OptionalUserAction.async { implicit request =>
     val format = "eac"
     val params = request.queryString.filterKeys(_ == "lang")
-    userBackend.stream(s"${EntityType.HistoricalAgent}/$id/$format", params = params).map { case (head, body) =>
+    userBackend.stream(s"classes/${EntityType.HistoricalAgent}/$id/$format", params = params).map { case (head, body) =>
       Status(head.status)
         .chunked(body.andThen(Enumerator.eof))
         .withHeaders(head.headers.map(s => (s._1, s._2.head)).toSeq: _*)
