+++ modules/portal/app/controllers/portal/Repositories.scala
@@ -92,7 +92,7 @@ case class Repositories @Inject()(
   def export(id: String) = OptionalUserAction.async { implicit request =>
     val format = "eag" // Hardcoded for now!
     val params = request.queryString.filterKeys(_ == "lang")
-    userBackend.stream(s"${EntityType.Repository}/$id/$format", params = params).map { case (head, body) =>
+    userBackend.stream(s"classes/${EntityType.Repository}/$id/$format", params = params).map { case (head, body) =>
       Status(head.status)
         .chunked(body.andThen(Enumerator.eof))
         .withHeaders(head.headers.map(s => (s._1, s._2.head)).toSeq: _*)
